# docs/swagger.yaml
definitions:
...
  dto.LoginUser:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterUser:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  models.User:
    properties:
      created_at:
        description: custom column name, auto-set when inserting
        type: string
      email:
        type: string
      id:
        description: auto-increment by default
        type: integer
      name:
        type: string
      updated_at:
        description: auto-set when inserting/updating
        type: string
    required:
    - email
    - name
    type: object
info:
  contact: {}
  description: This is a Go REST API server for an events app.
  title: Go Events API server
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Logs in a user
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
      summary: Registers a new user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Enter your JWT token in the format Bearer **&lt;token&gt;**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"